remove default value for max_connections
the low default value for the maximum number of connections is already
conflicting with the postgres configuration on ubuntu versions for the
second time (18.04, now 19.04).

it seems wiser to let the postgres configuration of the OS decide the
default value for the maximum number of connections (it can still be
modified via the constructor).